requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]

pytest
pytest-playwright
